oimo_ohapi/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
celery[redis]==5.4.0
redis==5.0.8
sqlalchemy==2.0.35
psycopg[binary]==3.2.2
alembic==1.13.3
boto3==1.35.24
python-multipart==0.0.12
httpx==0.27.2
pillow==10.4.0
numpy==2.1.1
insightface==0.7.3
onnxruntime==1.19.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
